pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "corbell"
version = "0.2.0"
description = "Multi-repo architecture graph, spec generation, and architecture review for backend teams."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
  {name = "Corbell Contributors"}
]
keywords = ["architecture", "spec", "design-docs", "multi-repo", "embeddings", "LLM"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: Apache Software License",
  "Topic :: Software Development :: Documentation",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
  "typer>=0.12",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "gitpython>=3.1",
  "sentence-transformers>=3.0",
  "numpy>=1.26",
  "rich>=13.0",
  "questionary>=2.0",
]

[project.optional-dependencies]
neo4j    = ["neo4j>=5.0"]
chroma   = ["chromadb>=0.5"]
openai   = ["openai>=1.0"]
anthropic = ["anthropic>=0.25"]
notion   = ["notion-client>=2.0"]
linear   = ["requests>=2.31"]
dev      = [
  "pytest>=8.0",
  "pytest-cov",
  "pytest-asyncio",
  "respx",
  "httpx",
  "ruff",
  "mypy",
]

[project.scripts]
corbell = "corbell.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["corbell"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
